♻️ Refactor: seperate ChatRoomContent component (#118)
